web/requirements.txt
# Global
Django
psycopg2-binary
Pillow

# Django core
django-split-settings

# Development
django-debug-toolbar
isort
geocoder

# App specific
django-model-utils
django-phonenumber-field
django-autocomplete-light
social-auth-app-django
django-extra-views

# For frontend
django-bootstrap4